Add hyphen + space to span preserve characters (pre)
CleanupMacro will close up space after first hyphen in a phrase like
"pre- and post-war". Now the hyphen + space can be tagged and will be
preserved. This doesn't totally solve the problem, because it has to
be identified and tagged. In future will need to figure out how to
search for and avoid this as standard part of macro.