Add function to remove hidden text. Fixes #154. Fixes #119.
Character Styles macro was converting hidden text to regular text, and
would throw an error if the hidden text was in embedded notes. This
function just does a find and delete for hidden text. It returns
True/False, which isn't used for anything yet but we could notify the
user at some point if necessary. Adds function to Character Styles and
Cleanup macros.