Tagging Bookmaker line-fixing styles if they also have format
If use any "bookmaker tighten ()" format style, and the span
also has bold/ital/whatever formatting, will append the bkmkr
action to the end of the original style name. Currently only
works predictably if the whole span has both the bkmkr style
and the formatting. If you have the bkmkr style in a sub-span
of that formatting, will give wonky results.