Removing invalid field codes. Fixes #172.
Added to SharedMacros.StartupSettings. Prev. version just
searched for any Field Codes and tried to replace field with
text of its result. Turns out some kinds of Fields don't have
a result, so this was casuing an error. Now it checks for the
.Kind property, and if it's one that doesn't produce an result,
we just delete it. Also loops through all stories now too, so
we'll get the files in the notes as well.