Check for read-only before del. weird "toolbar" on PC
MacmillanGT/ThisDocument contains a thing to delete the
"toolbar" on MacGt--for some reason, it adds the macros
as the Ribbon tab AND as old-school "toolbar", but old-school
toolbars appear as an "add-ins" tab, so in effect everything
is duplicated. However, sometimes it sees the template as
read/write, which is fine, and other times as read-only, and
users are prompted to save and it's confusing. So this now
checks if it's read only, and if it is it just skips removing
the toolbar. Would like a better fix in the future, but for now
this will stop the notice from appearing.