Searching for bkmkr styles in use
The .InUse property of the Style object does not mean
"in use in document" (like in the Style Pane>Options sense)
it means "any custom style or modified built-in style"
which is silly but whatever. So just searching for .InUse =
True always returned all bkmkr styles in TagBkmkrCharStyles
function. So this adds a secondary search, using the
Selection.Find command to see if it's actually being used
before adding it to the array.