Converting all one-char vars to uppercase
The "Dim A As Long" variable situation is common
when making For loops, but it turns out that if you
declare "A" as a var in one module but "a" as a var
in another module, VBA will sometimes change case without
warning, and then might change back later. Anyway, it's
annoying and messing up my diffs so I just made all
of them uppercase everywhere. Going forward, only use
uppercase single-letter variables.