Adjusting style creation error handler
Char Styles macro now tags TX, TX1, and COTX1 for unstyled paras.
If the doc doesn't have that style, it is created. However, see
comments in code re: controlling what formatting is set under
what conditions. I'm not completely resolving it now because SOP
is to only run Char Styles on docs w/ most styles applied, so
needing to create the styles will be rare. Plus, in the near future
we'd like to develop some kind of way to create styles from an
external file, so we'll figure it out more successfully for that.